Zombies Implemented.
I added in the rest of the current Zombie animations that I have made, as well as created a basic navmesh. The zombie has health, and a death script, but since we dont have shooting it wont die just yet but the code is there. Models are all looking good, but some issues with the transform of the Orientation object when crouching or sliding make it so that the Hand and Pistol models get distorted.