All bugs fixed, Animations added.
I fixed the zombie AI. The issue with the navigation is because no new navmesh was created. We need to update this under "Window" > AI every single time we make changes to our level. Left Arm was missing normals in some areas, and AI did not act as it should. I increased the size of the zombie to match the height of the player, and added additional zombies to the scene. I temporarily removed crouching which is a fix for the weird transforms to the player arms and weapons.